requirements.txt
tensorflow >= 2.12.0
scikit-learn >= 1.2.2
keras >= 2.12.0
numpy >= 1.24.3
Pillow >= 9.5.0